fix the overflow bug

+++ src/main_FPGA.cpp
@@ -285,10 +285,10 @@ void DLL_EXPORT Configure_SetSamplingRate(
 
     pContainer[iContainer++] = SET_SPL_RATE;
 
-    uint8_t clk_cnt_max = (uint8_t) ((25000.0/nfreq)+0.5); //25000 kHz = 50/2 Mhz = 50000/2 kHz
+    uint32_t clk_cnt_max = (uint32_t) ((25000.0/nfreq)+0.5); //25000 kHz = 50/2 Mhz = 50000/2 kHz
     // +0.5 is for rounding. ref: http://stackoverflow.com/questions/9695329/c-how-to-round-a-double-to-an-int
 
-    pContainer[iContainer++] = clk_cnt_max/10;
+    pContainer[iContainer++] = (uint8_t)(clk_cnt_max/10);
 }
